// crates/opsml_server/opsml_ui/src/lib/scripts/types.ts
export enum ProfileType {
  SPC = "SPC",
  PSI = "PSI",
}
// ...
export interface SpcFeatureDriftProfile {
  id: string;
  center: number;
  one_ucl: number;
  one_lcl: number;
  two_ucl: number;
  two_lcl: number;
  three_ucl: number;
  three_lcl: number;
  timestamp: string;
}
// ...
export interface DriftValues {
  created_at: string[];
  values: number[];
}
// ...
export interface AlertMetrics {
  created_at: string[];
  acknowledged: number[];
  active: number[];
  alert_count: number[];
}